integrate pipeline functionality into llm_methods
