+++ llm_pipeline/llm_methods.py
@@ -1,12 +1,20 @@
-import pandas as pd
+import asyncio
+import hashlib
+import json
+import os
+from abc import ABC, abstractmethod
+from math import ceil
+from typing import Any, Callable, List
+
+import logging
+
 import numpy as np
-from typing import Callable, List, Any
+import pandas as pd
+from dotenv import load_dotenv
 from openai import OpenAI
 from sklearn.cluster import DBSCAN
-from dotenv import load_dotenv
-from math import ceil
-import os
-import logging
+
+from tools.utils import log_message, validate_input
 
 logger = logging.getLogger(__name__)
 
@@ -16,14 +24,22 @@ load_dotenv()
 # Utility: OpenAI embedding function (default implementation)
 # ------------------------------------------------------------------
 def openai_embedding_function(text: str) -> List[float]:
+    """Return an embedding vector for ``text``.
+
+    If ``OPENAI_API_KEY`` is available the function calls the OpenAI API,
+    otherwise it falls back to generating a deterministic pseudo embedding so
+    that the rest of the pipeline can operate without network access.
     """
-    Generate an embedding vector for the given text using OpenAI's embeddings model.
-    """
-    client = OpenAI()
+    api_key = os.getenv("OPENAI_API_KEY")
+    if not api_key:
+        digest = hashlib.sha256(text.encode("utf-8")).digest()
+        return [b / 255 for b in digest[:32]]
+
+    client = OpenAI(api_key=api_key)
     try:
         response = client.embeddings.create(
-            input=f"{text}",
-            model="text-embedding-3-small"
+            input=text,
+            model="text-embedding-3-small",
         )
         return response.data[0].embedding
     except Exception as e:  # pragma: no cover - network errors
@@ -372,4 +388,383 @@ class DataFrameProcessor:
 
     def get_df(self) -> pd.DataFrame:
         return self.df
+
+
+###############################################################################
+# Pipeline style classes migrated from pipeline module
+###############################################################################
+
+
+class PipelineStep(ABC):
+    """Abstract base class for a pipeline processing step."""
+
+    @abstractmethod
+    def process(self, df: pd.DataFrame) -> pd.DataFrame:
+        """Process ``df`` and return the updated DataFrame."""
+        raise NotImplementedError
+
+
+class LLMCallStep(PipelineStep):
+    """Call an LLM for each record with optional caching."""
+
+    def __init__(
+        self,
+        prompt_template: str,
+        output_key: str = "llm_output",
+        fields: List[str] | None = None,
+        cache_path: str | None = None,
+    ) -> None:
+        self.prompt_template = prompt_template
+        self.output_key = output_key
+        self.fields = fields
+        self.cache_path = cache_path
+        self.cache: dict[str, str] = {}
+        if cache_path and os.path.exists(cache_path):
+            try:
+                with open(cache_path, "r", encoding="utf-8") as f:
+                    self.cache = json.load(f)
+            except Exception:
+                self.cache = {}
+        self.system_prompt = (
+            "You are evaluating records. The user will submit a record to be evaluated, and you should respond with only the evaluation. \n"
+            "Respond only with the evaluation, and no other terms or formatting.\n\n"
+            "The requested evaluation is:"
+        )
+        api_key = os.getenv("OPENAI_API_KEY")
+        self.client = OpenAI(api_key=api_key) if api_key else None
+
+    def _get_llm_response(self, prompt: str) -> str:
+        if prompt in self.cache:
+            return self.cache[prompt]
+        if self.client is None:
+            response = "yes" if "ui" in prompt.lower() else "no"
+        else:
+            chat_completion = self.client.chat.completions.create(
+                model="gpt-4o-mini",
+                messages=[
+                    {"role": "system", "content": self.system_prompt},
+                    {"role": "user", "content": prompt},
+                ],
+            )
+            response = chat_completion.choices[0].message.content.strip()
+        self.cache[prompt] = response
+        if self.cache_path:
+            try:
+                with open(self.cache_path, "w", encoding="utf-8") as f:
+                    json.dump(self.cache, f)
+            except Exception:
+                pass
+        return response
+
+    def process(self, df: pd.DataFrame) -> pd.DataFrame:
+        def create_record_details(row: pd.Series) -> str:
+            details = {field: row[field] for field in self.fields} if self.fields else row.to_dict()
+            return "\n".join(f"{key}: {value}" for key, value in details.items())
+
+        df = df.copy()
+        df["record_details"] = df.apply(create_record_details, axis=1)
+        df["prompt"] = df["record_details"].apply(
+            lambda details: self.prompt_template.format(record_details=details)
+        )
+        df[self.output_key] = df["prompt"].apply(self._get_llm_response)
+        df.drop(columns=["record_details", "prompt"], inplace=True)
+        return df
+
+    async def process_async(self, df: pd.DataFrame) -> pd.DataFrame:
+        df = df.copy()
+
+        def create_record_details(row: pd.Series) -> str:
+            details = {field: row[field] for field in self.fields} if self.fields else row.to_dict()
+            return "\n".join(f"{key}: {value}" for key, value in details.items())
+
+        df["record_details"] = df.apply(create_record_details, axis=1)
+        df["prompt"] = df["record_details"].apply(
+            lambda details: self.prompt_template.format(record_details=details)
+        )
+
+        async def call(prompt: str) -> str:
+            return await asyncio.to_thread(self._get_llm_response, prompt)
+
+        responses = await asyncio.gather(*(call(p) for p in df["prompt"].tolist()))
+        df[self.output_key] = responses
+        df.drop(columns=["record_details", "prompt"], inplace=True)
+        return df
+
+
+class FixedProcessingStep(PipelineStep):
+    """A processing step that applies a fixed function."""
+
+    def __init__(
+        self,
+        process_function: Callable[[pd.DataFrame], pd.Series],
+        output_key: str,
+    ) -> None:
+        self.process_function = process_function
+        self.output_key = output_key
+
+    def process(self, df: pd.DataFrame) -> pd.DataFrame:
+        df[self.output_key] = self.process_function(df)
+        return df
+
+
+class FilterStep(PipelineStep):
+    """Filter rows in the DataFrame based on ``filter_function``."""
+
+    def __init__(self, filter_function: Callable[[pd.DataFrame], pd.Series]) -> None:
+        self.filter_function = filter_function
+
+    def process(self, df: pd.DataFrame) -> pd.DataFrame:
+        return df.loc[self.filter_function(df)]
+
+
+class GenerateEmbeddingsStep(PipelineStep):
+    """Generate embeddings for each record."""
+
+    def __init__(
+        self,
+        embedding_function: Callable[[str], List[float]] = openai_embedding_function,
+        output_key: str = "embedding",
+        fields: List[str] | None = None,
+    ) -> None:
+        self.embedding_function = embedding_function
+        self.output_key = output_key
+        self.fields = fields
+
+    def process(self, df: pd.DataFrame) -> pd.DataFrame:
+        def get_text(row: pd.Series) -> str:
+            if self.fields:
+                return " ".join(str(row[field]) for field in self.fields)
+            return " ".join(row.astype(str))
+
+        df[self.output_key] = df.apply(lambda row: self.embedding_function(get_text(row)), axis=1)
+        return df
+
+
+class kNNFilterStep(PipelineStep):
+    """Filter to the top ``k`` records most similar to ``query``."""
+
+    def __init__(
+        self,
+        query: str,
+        k: int,
+        embedding_function: Callable[[str], List[float]] = openai_embedding_function,
+        embedding_column: str = "embedding",
+    ) -> None:
+        self.query = query
+        self.k = k
+        self.embedding_function = embedding_function
+        self.embedding_column = embedding_column
+
+    def _cosine_similarity(self, vec_a: List[float], vec_b: List[float]) -> float:
+        a = np.array(vec_a)
+        b = np.array(vec_b)
+        norm_a = np.linalg.norm(a)
+        norm_b = np.linalg.norm(b)
+        if norm_a == 0 or norm_b == 0:
+            return 0.0
+        return float(np.dot(a, b) / (norm_a * norm_b))
+
+    def process(self, df: pd.DataFrame) -> pd.DataFrame:
+        query_embedding = self.embedding_function(self.query)
+        df["similarity"] = df[self.embedding_column].apply(
+            lambda emb: self._cosine_similarity(emb, query_embedding)
+        )
+        df_sorted = df.sort_values(by="similarity", ascending=False)
+        result_df = df_sorted.head(self.k).copy()
+        result_df.drop(columns=["similarity"], inplace=True)
+        return result_df
+
+
+class LLMCallWithDataFrame:
+    """Call an LLM with the entire DataFrame combined into one prompt."""
+
+    def __init__(self, prompt_template: str, fields: List[str] | None = None) -> None:
+        self.prompt_template = prompt_template
+        self.fields = fields
+        self.system_prompt = (
+            "You are processing a user request against a set of records.  Please respond to the request as directed, without any additional comments or text.\n\n"
+        )
+        api_key = os.getenv("OPENAI_API_KEY")
+        self.client = OpenAI(api_key=api_key) if api_key else None
+
+    def create_prompt(self, df: pd.DataFrame) -> str:
+        def create_record_details(row: pd.Series) -> str:
+            details = {field: row[field] for field in self.fields} if self.fields else row.to_dict()
+            return "\n".join(f"{key}: {value}" for key, value in details.items())
+
+        record_details = df.apply(create_record_details, axis=1).tolist()
+        combined = "\n\n".join(f"# new record\n{details}" for details in record_details)
+        return self.prompt_template.format(record_details=combined)
+
+    def call_llm(self, df: pd.DataFrame) -> str:
+        prompt = self.create_prompt(df)
+        if self.client is None:
+            return "LLM unavailable"
+        chat_completion = self.client.chat.completions.create(
+            model="gpt-4o-mini",
+            messages=[
+                {"role": "system", "content": self.system_prompt},
+                {"role": "user", "content": prompt},
+            ],
+        )
+        return chat_completion.choices[0].message.content.strip()
+
+
+class AgenticGoalStep(PipelineStep):
+    """An agentic step that attempts to accomplish a goal using tools."""
+
+    def __init__(
+        self,
+        goal: str,
+        *,
+        per_row: bool = False,
+        guidance: str = "",
+        tools: dict | None = None,
+        output_key: str = "agent_output",
+        use_mcp: bool = False,
+        mcp_servers: List[str] | None = None,
+        discover_mcp: bool = True,
+        max_steps: int = 3,
+    ) -> None:
+        self.goal = goal
+        self.per_row = per_row
+        self.guidance = guidance
+        self.tools = tools or {}
+        if use_mcp:
+            try:
+                from mcp import MCP_TOOLS
+
+                self.tools.update(MCP_TOOLS)
+            except Exception:
+                pass
+        self.mcp_servers = mcp_servers or []
+        self.discover_mcp = discover_mcp
+        for i, url in enumerate(self.mcp_servers):
+            if discover_mcp:
+                try:
+                    from mcp import discover_remote_tools
+
+                    discovered = discover_remote_tools(url)
+                    if discovered:
+                        self.tools.update(discovered)
+                        continue
+                except Exception:
+                    pass
+            self.tools[f"mcp_server_{i}"] = self._make_mcp_tool(url)
+        self.output_key = output_key
+        self.max_steps = max_steps
+        api_key = os.getenv("OPENAI_API_KEY")
+        self.client = OpenAI(api_key=api_key) if api_key else None
+
+    def _make_mcp_tool(
+        self, url: str, name: str | None = None, desc: str | None = None
+    ) -> Callable[[dict, dict | None], dict]:
+        """Return a wrapper that sends context to an MCP server or endpoint."""
+
+        endpoint = url if name is None else f"{url.rstrip('/')}/{name}"
+
+        def call(context: dict, args: dict | None = None) -> dict:
+            try:
+                import requests
+
+                payload = {"context": context, "args": args or {}}
+                resp = requests.post(endpoint, json=payload, timeout=10)
+                resp.raise_for_status()
+                return resp.json()
+            except Exception as exc:  # pragma: no cover - network errors
+                return {"error": str(exc)}
+
+        call.__doc__ = desc
+
+        return call
+
+    def _agent_prompt(self, context: dict, history: list) -> str:
+        tool_desc = "\n".join(
+            f"- {name}: {func.__doc__ or 'no description'}" for name, func in self.tools.items()
+        )
+        hist = "\n".join(f"Step {i+1} result: {h}" for i, h in enumerate(history))
+        return (
+            f"Goal: {self.goal}\n{self.guidance}\nContext: {context}\n{hist}\n"
+            f"Available tools:\n{tool_desc}\n"
+            'Respond with JSON either {"tool": <name>, "args": <args>} '
+            'to invoke a tool or {"answer": <result>} to answer directly.'
+        )
+
+    def _run_agent(self, context: dict) -> str:
+        history: list = []
+
+        if self.client is None:
+            result = context
+            for func in self.tools.values():
+                result = func(result, None)
+                history.append(result)
+            return result
+
+        for _ in range(self.max_steps):
+            prompt = self._agent_prompt(context, history)
+            chat_completion = self.client.chat.completions.create(
+                model="gpt-4o-mini",
+                messages=[{"role": "user", "content": prompt}],
+            )
+            msg = chat_completion.choices[0].message.content.strip()
+            try:
+                data = json.loads(msg)
+            except json.JSONDecodeError:
+                return msg
+            if "tool" in data and data["tool"] in self.tools:
+                tool_output = self.tools[data["tool"]](context, data.get("args"))
+                history.append(tool_output)
+                context["last_tool_output"] = tool_output
+                continue
+            return data.get("answer", msg)
+        return str(history[-1]) if history else ""
+
+    def process(self, df: pd.DataFrame) -> pd.DataFrame:
+        if self.per_row:
+            df[self.output_key] = df.apply(lambda row: self._run_agent(row.to_dict()), axis=1)
+            return df
+        self.result = self._run_agent(df.to_dict(orient="records"))
+        return df
+
+
+class SummarizationStep(PipelineStep):
+    """Generate a summary for the entire DataFrame."""
+
+    def __init__(self, prompt_template: str, fields: List[str] | None = None) -> None:
+        self.llm = LLMCallWithDataFrame(prompt_template, fields)
+        self.summary: str = ""
+
+    def process(self, df: pd.DataFrame) -> pd.DataFrame:
+        self.summary = self.llm.call_llm(df)
+        return df
+
+
+class DataPipeline:
+    """Run a sequence of ``PipelineStep`` objects over a DataFrame."""
+
+    def __init__(self, steps: List[PipelineStep]):
+        self.steps = steps
+
+    def run(self, df: pd.DataFrame) -> pd.DataFrame:
+        validate_input(df, pd.DataFrame)
+        for step in self.steps:
+            log_message(f"Running step {step.__class__.__name__}")
+            df = step.process(df)
+            validate_input(df, pd.DataFrame)
+        return df
+
+
+class AsyncDataPipeline(DataPipeline):
+    """Run pipeline steps asynchronously when possible."""
+
+    async def run(self, df: pd.DataFrame) -> pd.DataFrame:  # type: ignore[override]
+        validate_input(df, pd.DataFrame)
+        for step in self.steps:
+            log_message(f"Running step {step.__class__.__name__} asynchronously")
+            if hasattr(step, "process_async"):
+                df = await step.process_async(df)
+            else:
+                df = await asyncio.to_thread(step.process, df)
+            validate_input(df, pd.DataFrame)
+        return df
